fix(CL-07): Product.price is Numeric; Decimal through the pricing path
Product.price was Float while every other money column — order totals,
line unit prices, payment amounts, store delivery fees — is Numeric(10, 2).
Checkout multiplied that float by a quantity and stored it into Numeric.

- migration f6b7c8d9e0a1: products.price Float -> Numeric(10, 2).
- order_service prices entirely in Decimal (subtotal, fee, total, cart
  total). float() happens only in serialize_quote and the checkout
  response — the JSON boundary.
- product list/detail serializers wrap price in float() (jsonify cannot
  emit Decimal).
- test_checkout_total_is_exact_not_float_drifted: 0.10 x 3 comes back as
  0.30, not 0.30000000000000004, from both /orders/preview and /orders,
  and persists as Decimal("0.30").

pytest: 49 passed.

+++ app/models/product.py
@@ -17,7 +17,8 @@ class Product(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(120), nullable=False)
     description = db.Column(db.Text, nullable=True)
-    price = db.Column(db.Float, nullable=False)
+    # Money is Numeric(10, 2) everywhere — never Float (CL-07).
+    price = db.Column(db.Numeric(10, 2), nullable=False)
     stock = db.Column(db.Integer, default=0)
 
     store_id = db.Column(db.Integer, db.ForeignKey("stores.id"),
